move all inline html events to event-listeners.js
this should make refactoring of commonly called functions easier
