+++ Fumen Utils_files/modified-glueFumen.js
@@ -1,4 +1,5 @@
 const {decoder, encoder, Field} = require('tetris-fumen');
+import { getDelimiter, LineTerminator } from './global-utils.js'
 
 const rowLen = 10;
 
@@ -201,18 +202,11 @@ function findRemainingPieces(field){
     return piecesFound;
 }
 
-function glueFumen(customInput=document.getElementById('input').value, removeLineClearBool=true){
-    var fumenCodes = [];
-	customInput = customInput.split("\t")
-
-    if(typeof customInput != 'object'){
-        throw new Error("Custom input for glueFumen must be an array")
-    }
-
-    for(let rawInput of customInput){
-        fumenCodes.push(...rawInput.split(/\s/));
-    }
-
+export default function glueFumen(){
+    var removeLineClearBool = true //always set to true
+    var input = document.getElementById('input').value.replace(/[Ddm]115@/gm,'v115@')
+	var fumenCodes = input.trim().split(LineTerminator)
+    
     var allPiecesArr = [];
     var allFumens = [];
     var fumenIssues = 0;
@@ -261,9 +255,6 @@ function glueFumen(customInput=document.getElementById('input').value, removeLin
         console.log("Warning: " + fumenIssues + " fumens couldn't be glued");
     }
 
-	let output = allFumens.join(delimiter);
+	let output = allFumens.join(getDelimiter());
 	document.getElementById('output').value = output;
 }
-
-
-
